successfully implemented rust multi-threading; added new routes
